pyproject.toml
[tool.poetry]
name = "moe-gthr-auth-server"
version = "0.10.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "LICENCE"
readme = "readme.org"
packages = [{include = "moe_gthr_auth_server"}]

[tool.poetry.dependencies]
python = "^3.11"
Flask-SQLAlchemy = "^3.0.5"
schema = "^0.7.5"
gunicorn = "^21.2.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.283"
pytest = "^7.4.0"
black = "^23.7.0"
requests = "^2.31.0"
pre-commit = "^3.3.3"
ruff-lsp = "^0.0.37"

[tool.black]
python = "^3.11"
line-length = 92

[tool.ruff]
line-length = 92

[tool.pytest.ini_options]
log_format = "%(asctime)s %(levelname)s %(message)s"
log_date_format = "%Y-%m-%d %H:%M:%S"
log_cli = 1
log_cli_level = "DEBUG"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
